requirements.txt
requests
pytest
allure-pytest
curlify
selene == 2.0.0b17
python-dotenv
selenium